split out TCM monad & related stuff into a separate module
