+++ disco.cabal
@@ -147,6 +147,7 @@ library
                        Disco.Typecheck.Constraints
                        Disco.Typecheck.Erase
                        Disco.Typecheck.Graph
+                       Disco.Typecheck.Monad
                        Disco.Typecheck.Solve
                        Disco.Typecheck.Unify
                        Disco.Types
