+++ src/Disco/Desugar.hs
@@ -43,7 +43,7 @@ import           Disco.AST.Desugared
 import           Disco.AST.Surface
 import           Disco.AST.Typed
 import           Disco.Syntax.Operators
-import           Disco.Typecheck
+import           Disco.Typecheck.Monad
 import           Disco.Types
 
 ------------------------------------------------------------
@@ -249,7 +249,7 @@ desugarTerm (ATChain _ t1 links)  = desugarTerm $ expandChain t1 links
 
 desugarTerm (ATContainer ty c es mell) = desugarContainer ty c es mell
 
-desugarTerm (ATContainerComp ty ListContainer bqt) = do
+desugarTerm (ATContainerComp _ ListContainer bqt) = do
   (qs, t) <- unbind bqt
   desugarComp t qs
